Added Buttons w/ Event Handlers:
Added the 3 action buttons from App design to the HOH menu page.
Now, you can log in with bar@example.com : world, and click these
super legit buttons.  The functions that handle the button taps run when the event is triggered, but need to be wired up to the programming logic and support for Skype, and the database/client/Server API.