Added Hearing Tool Functionality:
Now you can record from the hearing tool, and type text to
facilitate conversation with a heard-of-hearing person!  Uses
animations to make the messages slide in, and correctly labels them.

Also cleaned up some whitepsace with LoginActivity.java and tried
to make it so the keyboard doesn't automatically come up using a trick
from a previous Android project.