Added Login/Signup through API:
I have added all the functionality to log in, requesting pass from API "getPassword?userEmail=", then veritying that the passwords match on the client side.
I have added all the functionality to sign up, generating a new hash/salt, and sending that over to the DB through the API with new email, and new password.

What doesn't work?

I need a decent way to get the user type back from that getPassword method on the API.  that way I can correcly navigate to the right splash page, and won't have to make a separate call to getUser when I don't need to.